+++ items.py
@@ -6,27 +6,40 @@
 # Description: Creates classes and sub-classes for the different
 # types of items.
 
+
+def item_usable(item):
+    """Checks to see if an item is still usable."""
+    if isinstance(item, (Heal, Weapon)):
+        return item.uses != 0
+    else:
+        return True
+
+
 class Item():
-    """A parent class for the characteristic of inventory items"""
+    """A parent class for the characteristic of inventory items."""
     def __init__(self, name, description, uses):
+        """Sets variables up for item name, description and uses."""
         self.name = name
         self.description = description
         self.uses = uses
 
     def __str__(self):
-        """Returns the name of each item"""
+        """Returns the name of each item."""
         return self.name
-    
+
     def __eq__(self, other):
+        """Called when comparing two items, returns if the items have the same
+        name.
+        """
         return other and self.name == other.name
 
 
 class Weapon(Item):
-    """Class for the weapon type of item used to deal damage"""
+    """Class for the weapon type of item used to deal damage."""
     def __init__(self, name, description, uses, damage):
         super().__init__(name, description, uses)
         self.damage = damage
-    
+
     @staticmethod
     def Knife():
         """Method for the knife weapon with the name, description, uses, and
@@ -38,7 +51,7 @@ class Weapon(Item):
             damage=25,
             uses=1000000
         )
-    
+
     @staticmethod
     def Chainsaw():
         """Method for the chainsaw weapon with the name, description, uses, and
@@ -48,9 +61,9 @@ class Weapon(Item):
             name="chainsaw",
             description="High Damage Weapon doing 125 damage",
             damage=125,
-            uses=10
+            uses=2
         )
-    
+
     @staticmethod
     def Axe():
         """Method for the axe weapon with the name, description, uses, and
@@ -62,7 +75,7 @@ class Weapon(Item):
             damage=90,
             uses=2
         )
-    
+
     @staticmethod
     def Sword():
         """Method for the sword weapon with the name, description, uses, and
@@ -72,7 +85,7 @@ class Weapon(Item):
             name="sword",
             description="Large Weapon that does 75 damage",
             damage=75,
-            uses=10
+            uses=8
         )
 
 
@@ -81,11 +94,11 @@ class Heal(Item):
     def __init__(self, name, description, uses, add_health):
         super().__init__(name, description, uses)
         self.add_health = add_health
-    
+
     @staticmethod
     def Medkit():
-        """Method for the medkit item with name, description, uses, and the 
-        amount of health that will be added when used
+        """Method for the medkit item with name, description, uses, and the
+        amount of health that will be added when used.
         """
         return Heal(
             name="medkit",
@@ -93,15 +106,15 @@ class Heal(Item):
             uses=1,
             add_health=100
         )
-    
+
     @staticmethod
     def Bandages():
-        """Method for a bandage item with name, description, uses, and the 
-        amount of health that will be added when used
+        """Method for a bandage item with name, description, uses, and the
+        amount of health that will be added when used.
         """
         return Heal(
             name="bandages",
             description="Small healable item, restoring 50 health",
-            uses=1,
+            uses=2,
             add_health=50
         )