+++ main.py
@@ -4,7 +4,7 @@
 # Date last modified: 2020-10-25
 # Name: Aarsh Shah
 # Description: Imports modules/ classes and creates a map with different game
-# tiles with user inputted movement.
+# tiles with user inputted movement, attacking, healing and more.
 
 import random
 from copy import deepcopy
@@ -17,20 +17,13 @@ import menu
 from items import *
 from map import Maps
 from player import Player
-from villian import Immortalman, Villian
-
-
-def item_usable(item):
-    """Checks to see if an item is still usable."""
-    if isinstance(item, (Heal, Weapon)):
-        return item.uses != 0
-    else:
-        return True
+from villian import Villian
 
 
 class Game:
     """A class that allows the game to be run taking user inputs to move your
-    character, pick up items, attack enemies and run away."""
+    character, pick up items, attack enemies and run away.
+    """
     def __init__(self):
         """Sets variables up for the game, including the map,
         pending messages, the default settings.
@@ -67,7 +60,8 @@ class Game:
 
     def mainmenu(self):
         """Allows for user inputs as the map is printed and reacts according to
-        the users inputs."""
+        the users inputs.
+        """
         map_copy = deepcopy(self.current_map)
         # Sets the marker to indicate the player as "You".
         map_copy[self.pos[0]][self.pos[1]] = "You"
@@ -121,8 +115,7 @@ Left, Right, Up, or Down?
                 self.change_map(hotspot["new_map"])
 
     def positioning(self):
-        """Method to store what enemy
-        is at your current location"""
+        """Method to store what enemy is at your current location."""
         # Sets the positon to x and y coordinates.
         self.y, self.x = self.pos
         # Variable to store what enemy is in battle.
@@ -154,7 +147,8 @@ Left, Right, Up, or Down?
 
     def menu1(self):
         """Creates the menu that is presented when the player encounters
-        an enemy."""
+        an enemy.
+        """
         self.positioning()
         while 1:
             # Variable that stores the input of the players next action
@@ -183,12 +177,15 @@ What do you want to do?
 
     def equip(self):
         """Method to allow the player to equip either a healable or weapon
-        item."""
+        item.
+        """
         self.positioning()
         # A while loop that runs as long as the enemy is alive.
         while self.enemy.alive():
+            # Checks to see if the items in the inventory still have uses.
             self.player.inventory = list(
                 filter(item_usable, self.player.inventory))
+            # Asks the user what they want to equip from their inventory.
             print("What do you want equip?\n")
             print(self.player.inventory_info())
             # Variable stores the user input for what item they want to equip.
@@ -198,7 +195,7 @@ What do you want to do?
                 self.equipped_item = next(item
                                           for item in self.player.inventory
                                           if item.name == option)
-                print(f"You equipped a{str(self.equipped_item)}")
+                print(f"You equipped {str(self.equipped_item)}")
                 # Checks to see if the chosen item is a weapon.
                 if isinstance(self.equipped_item, Weapon):
                     self.attack()
@@ -209,27 +206,29 @@ What do you want to do?
                 if not self.player.alive():
                     print("You Died!")
                     quit()
+                # Removes the enemy off the map once they've been killed.
                 if not self.enemy.alive():
                     self.enemy_map[self.y][self.x] = None
             elif option == "quit":
                 quit()
             else:
                 print("You don't have that item")
-        # Returns a statement if you are successful in killing the enemy.
+        # Prints a statment saying that you've won if you killed the main boss.
         if self.enemy.name == "Immortal Man":
             print("""\
 You killed the Immortal Man!
 He was stashing the radio in his house.
 YOU HAVE ESCPAED, CONGRATS""")
             quit()
+        # Returns a statement if you are successful in killing the enemy.
         else:
             return ("You killed the enemy!")
-
 
     def heal(self):
         """Allows the player to recover health and use a healable."""
         self.player.health += self.equipped_item.add_health
         self.equipped_item.uses -= 1
+        print(f"You now have {self.player.health} health")
 
     def attack(self):
         """Allows the player to attack using their equipped weapon."""